# apispec.json
{"openapi":"3.0.3","info":{"title":"My Climate Data API","version":"1.0","description":"This specification is a description of two APIs working on the climate dataset provided at https://www.kaggle.com/unitednations/international-greenhouse-gas-emissions:\nAPI 1 - [host]/countries\nExample link - https://measure-floor.herokuapp.com/countries\n\nAPI 2 - [host]/country/id?queries\nExample link - https://measure-floor.herokuapp.com/country/1?parameters=CO2,N2O,PFCS,HFCS,GHGS,NF3,CH4,SF6&startYear=1995&endYear=2010\n","x-logo":{"url":""}},"servers":[{"url":"https://measure-floor.herokuapp.com","description":"","x-last-modified":1635712520766}],"paths":{"/countries":{"summary":"Get all countries in the dataset (names, ids and their possible values for startYear and endYear)","get":{"tags":[],"responses":{"JSON Response":{"content":{"application/json":{"encoding":{"data":[{"id":"0","country_or_area":"Australia","minimum_year":1990,"maximum_year":2014},{"id":"1","country_or_area":"Austria","minimum_year":1990,"maximum_year":2014},{"id":"2","country_or_area":"Belarus","minimum_year":1990,"maximum_year":2014},{"id":"3","country_or_area":"Belgium","minimum_year":1990,"maximum_year":2014},{"id":"4","country_or_area":"Bulgaria","minimum_year":1990,"maximum_year":2014},{"id":"5","country_or_area":"Canada","minimum_year":1990,"maximum_year":2014},{"id":"6","country_or_area":"Croatia","minimum_year":1990,"maximum_year":2014},{"id":"7","country_or_area":"Cyprus","minimum_year":1990,"maximum_year":2014},{"id":"8","country_or_area":"Czech Republic","minimum_year":1990,"maximum_year":2014},{"id":"9","country_or_area":"Denmark","minimum_year":1990,"maximum_year":2014},{"id":"10","country_or_area":"Estonia","minimum_year":1990,"maximum_year":2014},{"id":"11","country_or_area":"European Union","minimum_year":1990,"maximum_year":2014},{"id":"12","country_or_area":"Finland","minimum_year":1990,"maximum_year":2014},{"id":"13","country_or_area":"France","minimum_year":1990,"maximum_year":2014},{"id":"14","country_or_area":"Germany","minimum_year":1990,"maximum_year":2014},{"id":"15","country_or_area":"Greece","minimum_year":1990,"maximum_year":2014},{"id":"16","country_or_area":"Hungary","minimum_year":1990,"maximum_year":2014},{"id":"17","country_or_area":"Iceland","minimum_year":1990,"maximum_year":2014},{"id":"18","country_or_area":"Ireland","minimum_year":1990,"maximum_year":2014},{"id":"19","country_or_area":"Italy","minimum_year":1990,"maximum_year":2014},{"id":"20","country_or_area":"Japan","minimum_year":1990,"maximum_year":2014},{"id":"21","country_or_area":"Latvia","minimum_year":1990,"maximum_year":2014},{"id":"22","country_or_area":"Liechtenstein","minimum_year":1990,"maximum_year":2014},{"id":"23","country_or_area":"Lithuania","minimum_year":1990,"maximum_year":2014},{"id":"24","country_or_area":"Luxembourg","minimum_year":1990,"maximum_year":2014},{"id":"25","country_or_area":"Malta","minimum_year":1990,"maximum_year":2014},{"id":"26","country_or_area":"Monaco","minimum_year":1990,"maximum_year":2013},{"id":"27","country_or_area":"Netherlands","minimum_year":1990,"maximum_year":2014},{"id":"28","country_or_area":"New Zealand","minimum_year":1990,"maximum_year":2014},{"id":"29","country_or_area":"Norway","minimum_year":1990,"maximum_year":2014},{"id":"30","country_or_area":"Poland","minimum_year":1990,"maximum_year":2014},{"id":"31","country_or_area":"Portugal","minimum_year":1990,"maximum_year":2014},{"id":"32","country_or_area":"Romania","minimum_year":1990,"maximum_year":2014},{"id":"33","country_or_area":"Russian Federation","minimum_year":1990,"maximum_year":2014},{"id":"34","country_or_area":"Slovakia","minimum_year":1990,"maximum_year":2014},{"id":"35","country_or_area":"Slovenia","minimum_year":1990,"maximum_year":2014},{"id":"36","country_or_area":"Spain","minimum_year":1990,"maximum_year":2014},{"id":"37","country_or_area":"Sweden","minimum_year":1990,"maximum_year":2014},{"id":"38","country_or_area":"Switzerland","minimum_year":1990,"maximum_year":2014},{"id":"39","country_or_area":"Turkey","minimum_year":1990,"maximum_year":2014},{"id":"40","country_or_area":"Ukraine","minimum_year":1990,"maximum_year":2014},{"id":"41","country_or_area":"United Kingdom","minimum_year":1990,"maximum_year":2014},{"id":"42","country_or_area":"United States of America","minimum_year":1990,"maximum_year":2014}]}}},"x-last-modified":1635713961945,"description":"JSON response with a list of countries in the dataset.","headers":{},"links":{}}}},"x-last-modified":1635712567458},"/country/{id}":{"summary":"Get data for single country by its id. Returns all values for the selected parameters between startYear and endYear.","get":{"tags":[],"responses":{"JSON Response":{"description":"JSON response with all the records which match the queries in the request","content":{"application/json":{"encoding":{"data":[{"id":"1","country_or_area":"Austria","year":2010,"value":"335.8706445","category":"sulphur_hexafluoride_sf6_emissions_in_kilotonne_co2_equivalent"},{"id":"1","country_or_area":"Austria","year":2009,"value":"341.6786318","category":"sulphur_hexafluoride_sf6_emissions_in_kilotonne_co2_equivalent"},{"id":"1","country_or_area":"Austria","year":2008,"value":"373.4344819","category":"sulphur_hexafluoride_sf6_emissions_in_kilotonne_co2_equivalent"},{"id":"1","country_or_area":"Austria","year":2007,"value":"367.0066569","category":"sulphur_hexafluoride_sf6_emissions_in_kilotonne_co2_equivalent"},{"id":"1","country_or_area":"Austria","year":2006,"value":"453.4624463","category":"sulphur_hexafluoride_sf6_emissions_in_kilotonne_co2_equivalent"},{"id":"1","country_or_area":"Austria","year":2005,"value":"493.6250143","category":"sulphur_hexafluoride_sf6_emissions_in_kilotonne_co2_equivalent"},{"id":"1","country_or_area":"Austria","year":2004,"value":"484.0066031","category":"sulphur_hexafluoride_sf6_emissions_in_kilotonne_co2_equivalent"},{"id":"1","country_or_area":"Austria","year":2003,"value":"549.4427452","category":"sulphur_hexafluoride_sf6_emissions_in_kilotonne_co2_equivalent"},{"id":"1","country_or_area":"Austria","year":2002,"value":"613.300884","category":"sulphur_hexafluoride_sf6_emissions_in_kilotonne_co2_equivalent"},{"id":"1","country_or_area":"Austria","year":2001,"value":"629.3257066","category":"sulphur_hexafluoride_sf6_emissions_in_kilotonne_co2_equivalent"},{"id":"1","country_or_area":"Austria","year":2000,"value":"574.528064","category":"sulphur_hexafluoride_sf6_emissions_in_kilotonne_co2_equivalent"},{"id":"1","country_or_area":"Austria","year":1999,"value":"676.368455","category":"sulphur_hexafluoride_sf6_emissions_in_kilotonne_co2_equivalent"},{"id":"1","country_or_area":"Austria","year":1998,"value":"869.8823666","category":"sulphur_hexafluoride_sf6_emissions_in_kilotonne_co2_equivalent"},{"id":"1","country_or_area":"Austria","year":1997,"value":"1086.396974","category":"sulphur_hexafluoride_sf6_emissions_in_kilotonne_co2_equivalent"},{"id":"1","country_or_area":"Austria","year":1996,"value":"1176.899675","category":"sulphur_hexafluoride_sf6_emissions_in_kilotonne_co2_equivalent"},{"id":"1","country_or_area":"Austria","year":1995,"value":"1100.107713","category":"sulphur_hexafluoride_sf6_emissions_in_kilotonne_co2_equivalent"}]}}},"x-last-modified":1635714241712,"headers":{},"links":{}}}},"parameters":[{"deprecated":false,"name":"id","description":"id of the country for which data has to be fetched. All countries endpoint can be used to get list of countries with their ids.","schema":{"type":"string"},"in":"path","required":true,"x-last-modified":1635713676433},{"deprecated":false,"name":"startYear","description":"Used to restrict data with year greater than or equal to the value provided as startYear","schema":{"type":"number"},"in":"query","required":false,"allowEmptyValue":false,"x-last-modified":1635713691807},{"deprecated":false,"name":"endYear","description":"Used to restrict data with year greater than or equal to the value provided as endYear","schema":{"type":"number"},"in":"query","required":false,"allowEmptyValue":false,"x-last-modified":1635713702111},{"deprecated":false,"name":"parameters","description":"Comma separated list of climate parameters that can be filtered in the output. Following parameters are allowed:\nCO2\nN2O\nPFCS\nHFCS\nGHGS\nNF3\nCH4\nSF6","schema":{"type":"string"},"in":"query","required":false,"allowEmptyValue":false,"x-last-modified":1635713727925}],"x-last-modified":1635713614546}},"components":{"securitySchemes":{},"schemas":{},"headers":{},"responses":{},"parameters":{}},"tags":[],"security":[]}